+++ src/Exception/PhoneNumberParseException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Mollie\Exception;
+
+/**
+ * At the time of writing this class forwards errors from NumberParseException in libphonenumber library
+ */
+class PhoneNumberParseException extends MollieException
+{
+}