+++ src/Repository/CountryRepository.php
@@ -35,10 +35,16 @@
 
 namespace Mollie\Repository;
 
+use Country;
 use Db;
 
-class CountryRepository
+final class CountryRepository extends AbstractRepository
 {
+    public function __construct()
+    {
+        parent::__construct(Country::class);
+    }
+
     public function getMethodCountryIds($methodId)
     {
         $sql = 'SELECT id_country FROM `' . _DB_PREFIX_ . 'mol_country` WHERE id_method = "' . pSQL($methodId) . '"';
@@ -84,7 +90,7 @@ class CountryRepository
 
     public function getExcludedCountryIds($methodId)
     {
-        $sql = 'SELECT id_country 
+        $sql = 'SELECT id_country
                     FROM `' . _DB_PREFIX_ . 'mol_excluded_country`
                     WHERE id_method = "' . pSQL($methodId) . '"';
 
@@ -112,7 +118,7 @@ class CountryRepository
         $response = true;
         foreach ($idCountries as $idCountry) {
             $allCountries = 0;
-            $sql = 'INSERT INTO `' . _DB_PREFIX_ . 'mol_excluded_country` (id_method, id_country, all_countries) 
+            $sql = 'INSERT INTO `' . _DB_PREFIX_ . 'mol_excluded_country` (id_method, id_country, all_countries)
                 VALUES (';
 
             if ($idCountry === '0') {