+++ vendorBuilder/scoper.inc.php
@@ -25,7 +25,7 @@ return [
                 'test_old',
                 'tests',
                 'Tests',
-                'vendor-bin',
+                'vendor-bin'
             ])
             ->in('vendor'),
         Finder::create()->append([